requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
transformers
sentencepiece
pandas
Pillow
python-multipart
